advanced-api-project/requirements.txt
Django==5.2.4
djangorestframework==3.14.0 